Part_12 Types
The compiler from previous part only has 'INT' as type, this part we aim to add one more type, CHAR, and also prepare for future type additions.

-------------------------------------------------

On x86-64, INT takes 4 bytes while CHAR is 1 byte. CHAR can be seen as an INT, but INT cannot be taken as CHAR. Hence if we encounter assignment like char j = 3333, we must give it a warning of type incompatible.

In types.c, we have such a function checking the compatibility. So far we need to check the compatibility when EVALUATE EXPRESSION, PRINT STATEMENTS and in ASSIGNMENTS.

When evaluating expressions, newly parsed expression tokens should have the compatible type with previous tokens. So in binexpr(), we need to check the compatibility.

When printing statements. So far we only print CHAR or INT, so actually no type compatibility checking is necessary. Here we just check if the expression to be printed has type compatible to INT.

When assigning values to variables, type checking is important. We allow CHAR type to be assigned to a INT variable, but not the other way around.

Hence, in type_compatible() function, if any one type is P_VOID (corresponds to VOID return type), it is not compatible with the other. If both are the same type, free to go. If an CHAR is checked with INT, CHAR can be widen. But if an INT is assigned to CHAR (only happen in assignment statements, that is the function of onlyright parameter), they are not compatible and that CHAR will not be widen.

-------------------------------------------------

We add a field "TYPE" into AST struct, we add 4 types in an enumerate: P_NONE, P_VOID (only with function), P_INT, P_CHAR. Types will be passed when making an AST node.

P_NONE value indicates that the AST node does not represent an expression and has no type. For example when the node's operation is A_GLUE, A_PRINT, A_IF, A_WHILE, A_FOR.

When building AST for assignment statements, expressions, printing statements, when we check the type compatibility between types from left and right sub-child, if one side needs to be widen, then that side of sub tree will be nested with a unary node with node type A_WIDEN.

-------------------------------------------------

When generating assembly, we need to consider the types while generating a global symbol (allocating how many bytes for that in memory), load /initialise it (load the variable's value to a register) and store it (store value from register to variable).

To compile and test:
gcc -o comp1 -g cg.c decl.c expr.c gen.c main.c misc.c scan.c stmt.c sym.c tree.c types.c

./comp1 input13

gcc -o out out.s

./out