Part_13 Functions Part 2
In this part we extend previous ability on Functions. Beforehand, we were able to declare a  function structure. This time we can make a function call with a single argument and return from a non-void function (only). We also add a new type: LONG (8 bytes), aside from our previous CHAR(1 byte) and INT (4 bytes)

-----------------------------------------------

At this stage, the BNF syntax for a function call is

  function_call: identifier '(' expression ')'   ;

Only one parameter is supported at this moment and that parameter is not really passed into function yet. We just want to show that we can now do the function call with that grammar:
e.g. fred(5)

According to the BNF above, parsing function calls is easy. We firstly parse the function name as the identifier, then (, then argument expression, then ).  The function call AST will be an unary node with node type A_FUNCCALL and function's return type as type.

We can now have expression like x = fred + jim and x = fred(5) + jim. Both are assignment statement but the first fred is identifier while the second is function call. We hence need to add a function call case when parsing assignment statement.

-----------------------------------------------

We then deal with result of function. A function can return a value or not. We firstly parse a return statement. Its BNF is like:

return_statement: 'return' '(' expression ')'  ;

The parsing is easy as well. We also have to check the compatibility between function's type and returned type. We made a global variable Functionid to record the function we are currently in. The return statement itself is a unary node with node type being A_RETURN and type being P_NONE.

We need to ensure that a non-void function return a value. Note that function's body are statements which will be glued together in a final compound statement tree. According to our glueing algorithm, the final statement will appear at the top right of the tree. Hence when parsing function declaration, we need to check the top right node has the type A_RETURN if function is not void.

Note that at this stage we do not support statement which solely calls a void function.

-----------------------------------------------

Since a new type, LONG has been added. Our method of check type compatibility is now based on the size of the type. See type_compatible().

-----------------------------------------------

At last, assembly code is generated for calling and returning from a function. In x64, calling a function requires loading the (first) argument into %rdi, returned value from a function can be read from %rax. We made corresponding changes when moving data between registers and variables with respect to CHAR, INT and LONG.

-----------------------------------------------

To compile:

gcc -o comp1 -g cg.c decl.c expr.c gen.c main.c misc.c scan.c stmt.c sym.c tree.c types.c

./comp1 input14

gcc -o out out.s lib/printint.c (link assembly output with lib/printint.c)

./out